Kill DerivedStream.
