+++ build.sbt
@@ -13,7 +13,6 @@ lazy val root = (project in file(".")).
     libraryDependencies += "org.apache.parquet" % "parquet-avro" % "1.7.0",
     libraryDependencies += "com.github.seratch" %% "awscala" % "0.3.+",
     libraryDependencies += "net.sandrogrzicic" %% "scalabuff-runtime" % "1.4.0",
-    libraryDependencies += "com.typesafe" % "config" % "1.2.1",
     libraryDependencies += "org.scalatest" %% "scalatest" % "2.2.4",
     libraryDependencies += "org.xerial.snappy" % "snappy-java" % "1.1.2",
     libraryDependencies += "org.json4s" %% "json4s-jackson" % "3.2.10",
