--- src/main/resources/application.conf
@@ -1,3 +1,0 @@
-app {
-    parquetBucket="telemetry-parquet"
-}
